Move tests to __tests__ folder and finish moving widgets to the main package
