+++ packages/drupal-admin-ui/package.json
@@ -16,10 +16,12 @@
     "@material-ui/icons": "^3.0.1",
     "deepmerge": "^2.1.1",
     "drupal-ui": "^0.1.0",
+    "downshift": "^2.0.16",
     "emotion": "^9.0.2",
     "history": "^4.7.2",
     "interweave": "^8.0.2",
     "jss": "^9.8.2",
+    "keycode": "^2.2.0",
     "makecancelable": "^1.0.0",
     "prop-types": "^15.6.1",
     "qs": "^6.5.1",
@@ -32,6 +34,7 @@
     "react-router": "^4.3.1",
     "react-router-dom": "^4.2.2",
     "react-router-redux": "^5.0.0-alpha.9",
+    "react-rte": "^0.16.1",
     "react-sticky": "^6.0.2",
     "redux": "^3.7.2",
     "redux-devtools-extension": "^2.13.2",
@@ -42,6 +45,8 @@
     "@storybook/addons": "^3.4.8",
     "@storybook/react": "^3.4.8",
     "chromedriver": "^2.40.0",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
     "mkdirp": "^0.5.1",
     "nightwatch": "^0.9.21",
     "react-axe": "^3.0.1",
