+++ packages/drupal-admin-ui/src/components/05_pages/Permissions/__tests__/Permissions.js
@@ -1,4 +1,4 @@
-import { filterPermissions } from './Permissions';
+import { filterPermissions } from '../Permissions';
 
 const permissions = [
   {
